chore(server): renaming ProgrammingLanguage
