+++ server/src/models/Project.ts
@@ -12,7 +12,7 @@ import {
 import { IsBoolean, IsDate, IsString } from 'class-validator';
 
 import { BaseModel } from 'models/base/BaseModel';
-import { ProgramingLanguage } from 'models/ProgramingLanguage';
+import { ProgrammingLanguage } from 'models/ProgrammingLanguage';
 import { FileProject } from 'models/FileProject';
 import { SandpackTemplate } from 'models/SandpackTemplate';
 import { Member } from 'models/Member';
@@ -39,9 +39,9 @@ export class Project extends BaseModel {
 	files: FileProject[];
 
 	@Column('varchar')
-	@Field(() => ProgramingLanguage)
-	@ManyToOne(() => ProgramingLanguage, (language) => language.id, { eager: true })
-	programmingLanguage: ProgramingLanguage;
+	@Field(() => ProgrammingLanguage)
+	@ManyToOne(() => ProgrammingLanguage, (language) => language.id, { eager: true })
+	programmingLanguage: ProgrammingLanguage;
 
 	@Column('varchar', { nullable: true })
 	@Field(() => SandpackTemplate, { nullable: true })
